prod_requirements.txt
db-dtypes==1.2.0
google-api-core==2.19.0
google-auth==2.30.0
google-cloud-bigquery==3.24.0
google-cloud-bigquery-storage==2.25.0
google-cloud-core==2.4.1
langdetect==1.0.9
loguru==0.7.2
networkx==3.2.1
pyarrow==16.1.0
pyyaml==6.0.1
requests==2.32.3
scikit-learn==1.5.1
tqdm==4.66.4
xmltodict==0.13.0
numpy==1.26.4
python-box==7.1.1
torch==2.3.1
pandas==2.2.2